Move mem_persister to testutil package
